requirements.txt
discord.py
quart
toml

